# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "API Documentation for the E-Commerce platform.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/admin/categories": {
            "get": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Get all categories (Admin)",
                "description": "Retrieve a list of all categories.",
                "operationId": "3fc11a81efbb03b2004237b1ca11fdb9",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Create a new category (Admin)",
                "description": "Add a new category with an optional image.",
                "operationId": "b78202d020b5c79443c4e8e74bcf50e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/categories/{id}": {
            "put": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Update a category (Admin)",
                "description": "Update an existing category.",
                "operationId": "0618d8c185e1790c0729b548cc5ff4d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Delete a category (Admin)",
                "description": "Delete a category from the database.",
                "operationId": "445038f33e36fda29d31166b5ea6e5c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/subcategories": {
            "post": {
                "tags": [
                    "Admin - Subcategories"
                ],
                "summary": "Create a new subcategory (Admin)",
                "description": "Add a new subcategory.",
                "operationId": "8db2d4ff908a186cda7c27f0fb879e24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/subcategories/{id}": {
            "put": {
                "tags": [
                    "Admin - Subcategories"
                ],
                "summary": "Update a subcategory (Admin)",
                "description": "Update an existing subcategory.",
                "operationId": "eb7cf94446abbcdba78d4f890d592085",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Subcategories"
                ],
                "summary": "Delete a subcategory (Admin)",
                "description": "Delete a subcategory from the database.",
                "operationId": "dc4f0386fa6ab6e78c610992b2db6d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subcategory ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subcategory deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}